requirements.txt
pyodbc
pandas
customtkinter
python-dotenv
sqlalchemy
openpyxl
pillow